Added test for content based fusion and refactored blending weight calculation
