+++ src/multiview_stitcher/_tests/test_weights.py
@@ -1,19 +1,9 @@
-import os
-import tempfile
 import warnings
 
 import dask.array as da
-import pytest
 import xarray as xr
 
-from multiview_stitcher import (
-    fusion,
-    io,
-    msi_utils,
-    sample_data,
-    spatial_image_utils,
-    weights,
-)
+from multiview_stitcher import fusion, io, sample_data, spatial_image_utils
 from multiview_stitcher.io import METADATA_TRANSFORM_KEY
 
 
@@ -47,7 +37,7 @@ def test_fuse_field():
     assert xfused.dtype == sims[0].dtype
 
 
-def test_fuse_sims_two_channels():
+def test_fuse_sims():
     sims = io.read_mosaic_image_into_list_of_spatial_xarrays(
         sample_data.get_mosaic_sample_data_path()
     )
@@ -79,64 +69,3 @@ def test_fuse_sims_two_channels():
         METADATA_TRANSFORM_KEY
         in spatial_image_utils.get_tranform_keys_from_sim(xfused)
     )
-
-
-@pytest.mark.parametrize(
-    "ndim, weights_method",
-    [
-        (2, None),
-        (2, weights.content_based),
-        (3, None),
-        (3, weights.content_based),
-    ],
-)
-def test_multi_view_fusion(ndim, weights_method):
-    nviews = 3
-
-    sims = [
-        sample_data.generate_tiled_dataset(
-            ndim=ndim,
-            overlap=0,
-            N_c=1,
-            N_t=1,
-            tile_size=20,
-            tiles_x=1,
-            tiles_y=1,
-            tiles_z=1,
-            spacing_x=1,
-            spacing_y=1,
-            spacing_z=1,
-        )[0]
-        for _ in range(nviews)
-    ]
-
-    # prepare assertion
-    for _, sim in enumerate(sims):
-        sim.data += 1
-
-    fused = fusion.fuse(
-        sims[:],
-        transform_key=METADATA_TRANSFORM_KEY,
-        weights_method=weights_method,
-        weights_method_kwargs={"sigma_1": 1, "sigma_2": 2}
-        if weights_method == weights.content_based
-        else None,
-    )
-
-    with tempfile.TemporaryDirectory() as tmpdir:
-        fused.data = da.to_zarr(
-            fused.data,
-            os.path.join(tmpdir, "fused_sim.zarr"),
-            overwrite=True,
-            return_stored=True,
-            compute=True,
-        )
-
-        mfused = msi_utils.get_msim_from_sim(fused, scale_factors=[])
-
-        fused_path = os.path.join(tmpdir, "fused.zarr")
-        mfused.to_zarr(fused_path, mode="w")
-
-        mfused = msi_utils.multiscale_spatial_image_from_zarr(fused_path)
-
-        assert fused.data.min() > 0
