+++ src/multiview_stitcher/fusion.py
@@ -276,7 +276,7 @@ def fuse_field(
 
         # calculate blending weights
         field_w = xr.zeros_like(sim)
-        field_w.data = get_smooth_border_weight_from_shape(
+        field_w.data = weights.get_smooth_border_weight_from_shape(
             sim.shape[-ndim:],
             chunks=sim.chunks,
             widths=blending_widths,
@@ -596,86 +596,6 @@ def calc_stack_properties_from_volume(volume, spacing):
     properties_dict["origin"] = origin
 
     return properties_dict
-
-
-def smooth_transition(x, x_offset=0.5, x_stretch=None, k=3):
-    """
-    Transform the distance from the border to a weight for blending.
-    """
-    # https://math.stackexchange.com/questions/1832177/sigmoid-function-with-fixed-bounds-and-variable-steepness-partially-solved
-
-    if x_stretch is None:
-        x_stretch = x_offset
-
-    xp = x
-    xp = xp - x_offset  # w is 0 at offset
-    xp = xp / x_stretch / 2.0  # w is +/-0.5 at offset +/- x_stretch
-
-    w = 1 - 1 / (1 + (1 / (xp + 0.5) - 1) ** (-k))
-
-    w[xp <= -0.5] = 0.0
-    w[xp >= 0.5] = 1.0
-
-    return w
-
-
-def get_smooth_border_weight_from_shape(shape, chunks, widths=None):
-    """
-    Get a weight image for blending that is 0 at the border and 1 at the center.
-    Transition widths can be specified for each dimension.
-
-    20230926: this adds too many small dask tasks, need to optimize
-
-    Possible better strategy:
-
-    Assume that only the border chunks of the input files need to be different
-    from da.ones.
-
-    OR: multiscale strategy
-    - first fuse smallest scale
-    - then go on with only those chunks that need to be filled in
-    - not clear if it's beneficial, as it's already known which chunks are needed
-
-    OR: Manually construct output array
-
-    [!!] OR: Use interpolation to calculate distance map
-    - reduce input to one value per chunk, i.e. chunksize of [1]*ndim
-    - use this to define which input chunk is required for fusion
-    - also use this to calculate weights at border
-      (only where above coarse interpolation step gave < 1)
-
-    """
-
-    ndim = len(shape)
-
-    # get distance to border for each dim
-
-    dim_dists = [
-        da.arange(shape[dim], chunks=chunks[dim]).astype(float)
-        for dim in range(ndim)
-    ]
-
-    dim_dists = [
-        da.min(da.abs(da.stack([dd, dd - shape[idim]])), axis=0)
-        for idim, dd in enumerate(dim_dists)
-    ]
-
-    dim_ws = [
-        smooth_transition(
-            dim_dists[dim], x_offset=widths[dim], x_stretch=widths[dim]
-        )
-        for dim in range(ndim)
-    ]
-
-    # get product of weights for each dim
-    w = da.ones(shape, chunks=chunks, dtype=float)
-    for dim in range(len(shape)):
-        tmp_dim_w = dim_ws[dim]
-        for _ in range(ndim - dim - 1):
-            tmp_dim_w = tmp_dim_w[:, None]
-        w *= tmp_dim_w
-
-    return w
 
 
 # from scipy import interpolate
